refactor: fix tasks
